rag-chatbot/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
langchain==0.1.0
langchain-community==0.0.10
langchain-openai==0.0.2
chromadb==0.4.18
sentence-transformers>=2.3.0
pypdf2==3.0.1
python-docx==1.1.0
python-multipart==0.0.6
jinja2==3.1.2
aiofiles==23.2.1
numpy>=1.26
pandas>=2.1
tiktoken>=0.5.2,<0.6.0
openai>=1.4.0
streamlit>=1.30.0
